reaname BankIdApi to  BankIdSOAPApi, move configs to the configuration package
