+++ member-service/src/main/java/com/hedvig/memberservice/services/bankid/BankIdAdapter.java
@@ -4,7 +4,7 @@ import com.hedvig.external.bankID.bankidTypes.CollectResponse;
 import com.hedvig.external.bankID.bankidTypes.OrderResponse;
 import java.io.UnsupportedEncodingException;
 
-public class BankIdAdapter implements BankIdApi {
+public class BankIdAdapter implements BankIdSOAPApi {
 
   private final com.hedvig.external.bankID.BankIdApi api;
