+++ member-service/src/main/java/com/hedvig/memberservice/services/bankid/BankIdSOAPApi.java
@@ -4,7 +4,7 @@ import com.hedvig.external.bankID.bankidTypes.CollectResponse;
 import com.hedvig.external.bankID.bankidTypes.OrderResponse;
 import java.io.UnsupportedEncodingException;
 
-public interface BankIdApi {
+public interface BankIdSOAPApi {
   OrderResponse auth();
 
   OrderResponse sign(String ssn, String message) throws UnsupportedEncodingException;
